requirements.txt
# Server Angel Dependencies
# Python 3.8+ required

# System monitoring
psutil>=5.9.0

# Environment variable management
python-dotenv>=1.0.0
